+++ integration/templates/nuxt-node/pages/pricing-table.vue
@@ -1,0 +1,3 @@
+<template>
+  <experimental_PricingTable />
+</template>