+++ packages/clerk-js/src/ui/elements/contexts/FlowMetadataContext.tsx
@@ -16,7 +16,8 @@ type FlowMetadata = {
     | 'blankCaptcha'
     | 'waitlist'
     | 'checkout'
-    | 'planDetails';
+    | 'planDetails'
+    | 'pricingTable';
   part?:
     | 'start'
     | 'emailCode'