+++ packages/clerk-js/src/ui/hooks/useCheckout.ts
@@ -1,16 +1,13 @@
+import type { ClerkAPIResponseError } from '@clerk/shared/error';
 import { useClerk, useOrganization, useUser } from '@clerk/shared/react';
-import type {
-  __experimental_CheckoutProps,
-  __experimental_CommerceCheckoutResource,
-  ClerkAPIError,
-} from '@clerk/types';
+import type { __experimental_CheckoutProps, __experimental_CommerceCheckoutResource } from '@clerk/types';
 import { useCallback, useEffect, useState } from 'react';
 
 import { useFetch } from './useFetch';
 
 export const useCheckout = (props: __experimental_CheckoutProps) => {
   const { planId, planPeriod, subscriberType = 'user' } = props;
-  const { __experimental_commerce } = useClerk();
+  const clerk = useClerk();
   const { organization } = useOrganization();
   const [currentCheckout, setCurrentCheckout] = useState<__experimental_CommerceCheckoutResource | null>(null);
 
@@ -20,9 +17,9 @@ export const useCheckout = (props: __experimental_CheckoutProps) => {
     isLoading,
     invalidate,
     revalidate,
-    error,
+    error: _error,
   } = useFetch(
-    __experimental_commerce?.__experimental_billing.startCheckout,
+    clerk.__experimental_commerce?.__experimental_billing.startCheckout,
     {
       planId,
       planPeriod,
@@ -31,6 +28,8 @@ export const useCheckout = (props: __experimental_CheckoutProps) => {
     undefined,
     `commerce-checkout-${user?.id}`,
   );
+
+  const error = _error as ClerkAPIResponseError | undefined;
 
   const updateCheckout = useCallback((newCheckout: __experimental_CommerceCheckoutResource) => {
     setCurrentCheckout(newCheckout);
@@ -48,6 +47,6 @@ export const useCheckout = (props: __experimental_CheckoutProps) => {
     isLoading,
     invalidate,
     revalidate,
-    isMissingPayerEmail: error?.errors.some((e: ClerkAPIError) => e.code === 'missing_payer_email'),
+    errors: error?.errors,
   };
 };